and {maybe}	R Documentation

Combine predicate functions to check if all are TRUE

Description

Combine predicate functions to check if all are TRUE

Usage

and(...)

Arguments

...
Predicate functions

Value

A predicate function

Examples

and(not_null, not_na)(1)
and(not_null, not_na)(NULL)

[Package maybe version 1.1.0 Index]