and_then2 {maybe}	R Documentation

Evaluate a binary maybe returning function on two maybe values

Description

Evaluate a binary maybe returning function on two maybe values

Usage

and_then2(.m1, .m2, .f, ...)

Arguments

.m1
A maybe value

.m2
A maybe value

.f
A binary maybe returning function to apply to the maybe values

...
Named arguments for the function .f

Value

A maybe value

Examples

and_then2(just(1), just(2), maybe(`+`))
and_then2(nothing(), just(2), maybe(`/`))

[Package maybe version 1.1.0 Index]