not_empty {maybe}	R Documentation

Check if a vector or data frame is empty

Description

Check if a vector or data frame is empty

Usage

not_empty(a)

Arguments

a
Object to check

Value

TRUE or FALSE

Examples

not_empty(integer())
not_empty(list())
not_empty(1:10)
not_empty(data.frame())
not_empty(data.frame(a = 1:10))

[Package maybe version 1.1.0 Index]